Refactor web scraping with multi-API support and improved flexibility
